! DataFile: fixed a crash du to missing mask files during migration step to v10

+++ package/clickpoints/DataFile.py
@@ -482,10 +482,12 @@ class DataFile:
             self.migrate_to_10_masks = masks
             self.db.execute_sql("CREATE TABLE `mask_tmp` (`id` INTEGER NOT NULL, `image_id` INTEGER NOT NULL, `data` BLOB NOT NULL, PRIMARY KEY(id), FOREIGN KEY(`image_id`) REFERENCES 'image' ( 'id' ) ON DELETE CASCADE)")
             for mask in masks:
-                im = np.asarray(PILImage.open(os.path.join(self.migrate_to_10_mask_path, mask[2])))
-                value = imageio.imwrite(imageio.RETURN_BYTES, im, format=".png")
-                value = peewee.binary_construct(value)
-                self.db.execute_sql("INSERT INTO mask_tmp VALUES (?, ?, ?)", [mask[0], mask[1], value])
+                tmp_maskpath = os.path.join(self.migrate_to_10_mask_path, mask[2])
+                if os.path.exists(tmp_maskpath):
+                    im = np.asarray(PILImage.open(tmp_maskpath))
+                    value = imageio.imwrite(imageio.RETURN_BYTES, im, format=".png")
+                    value = peewee.binary_construct(value)
+                    self.db.execute_sql("INSERT INTO mask_tmp VALUES (?, ?, ?)", [mask[0], mask[1], value])
             self.db.execute_sql("DROP TABLE mask")
             self.db.execute_sql("ALTER TABLE mask_tmp RENAME TO mask")
             self.db.execute_sql("PRAGMA foreign_keys = ON")
